+++ packages/index-page/src/components/editor-picks-mobile.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import Swipeable from 'react-swipeable'
 import PropTypes from 'prop-types'
 import postPropType from './prop-types/post'
 import styled from 'styled-components'
@@ -7,17 +6,15 @@ import styled from 'styled-components'
 import { getHref } from '../utils/getHref'
 import { truncate, breakPoints } from '../utils/style-utils'
 // components
-import MobileFlexSwipeable from './mobile-flex-swipeable'
 import CategoryName from './common-utils/category-name'
 import ImgWrapper from './common-utils/img-wrapper'
 import Section from './common-utils/section'
 import SectionName from './common-utils/section-name'
-import SwipableMixin from './common-utils/swipable-mixin'
 import TRLink from './common-utils/twreporter-link'
+import MobileSwiperList from './mobile-swiper-list'
 // constants
 import sectionStrings from '../constants/section-strings'
 import color from '../constants/color'
-import { itemWidthPct } from '../constants/mobile-mockup-specification'
 // @twreporter
 import { fontWeight, fontFamily } from '@twreporter/core/lib/constants/font'
 // lodash
@@ -25,8 +22,6 @@ import get from 'lodash/get'
 const _ = {
   get,
 }
-
-const mobileWidth = breakPoints.mobileMaxWidth
 
 const mockup = {
   img: {
@@ -41,9 +36,6 @@ const mockup = {
 const CarouselContainer = styled(Section)`
   padding-top: 0;
   background: ${color.white};
-  @media (min-width: ${breakPoints.tabletMinWidth}) {
-    display: none;
-  }
 `
 
 const TextFrame = styled.div`
@@ -64,7 +56,7 @@ const Title = styled.div`
   top: 50px;
   left: 50%;
   transform: translateX(-50%);
-  width: ${itemWidthPct}%;
+  width: 100%;
   font-size: 24px;
   font-weight: ${fontWeight.bold};
   font-family: ${fontFamily.title};
@@ -83,7 +75,7 @@ const Description = styled.div`
   left: 50%;
   font-size: 18px;
   transform: translateX(-50%);
-  width: ${itemWidthPct}%;
+  width: 100%;
   text-align: left;
   color: ${color.darkGray};
   ${truncate('absolute', 1.43, 3, 'white')};
@@ -93,28 +85,27 @@ const ImgFrame = styled.div`
   height: 186px;
 `
 
-const FadeInFadeOut = styled.div`
-  opacity: ${props => (props.isSelected ? '1' : '0')};
-  z-index: ${props => (props.isSelected ? '1' : '0')};
-  transition: 0.5s opacity linear;
-`
-
-class EditorPicksMobile extends SwipableMixin {
+class EditorPicksMobile extends React.Component {
   render() {
-    const onSwiping = (e, deltaX, deltaY, absX, absY) => {
-      // In order to avoid slightly vibrating while swiping left and right,
-      // we set a threshold to prevent scrolling.
-      // 10 is the the threshold value we set after manually testing.
-      if (absY < 10) {
-        e.preventDefault()
-      }
-    }
-    const ImageComp = post => {
+    const { data } = this.props
+    const flexItems = data.map(post => {
       const isExternal = _.get(post, 'is_external', false)
       const href = getHref(_.get(post, 'slug', 'error'), isExternal)
       const imgObj = _.get(post, 'hero_image') || _.get(post, 'og_image')
+      const key = _.get(post, 'id')
       return (
-        <TRLink href={href} redirect={isExternal}>
+        <TRLink href={href} redirect={isExternal} key={key}>
+          <TextFrame>
+            <Category>
+              {_.get(post, 'category_set[0].category.name', '')}
+            </Category>
+            <Title>
+              <TRLink href={href} redirect={isExternal}>
+                <TitleSpan>{_.get(post, 'title', '')}</TitleSpan>
+              </TRLink>
+            </Title>
+            <Description>{_.get(post, 'og_description', '')}</Description>
+          </TextFrame>
           <ImgFrame>
             <ImgWrapper
               alt={_.get(imgObj, 'description')}
@@ -129,62 +120,14 @@ class EditorPicksMobile extends SwipableMixin {
           </ImgFrame>
         </TRLink>
       )
-    }
-
-    const { data } = this.props
-    const flexItems = data.map(post => {
-      return (
-        <MobileFlexSwipeable.FlexItem
-          key={_.get(post, 'id')}
-          mobileWidth={mobileWidth}
-        >
-          {ImageComp(post)}
-        </MobileFlexSwipeable.FlexItem>
-      )
-    })
-
-    const textFrameContent = data.map((post, index) => {
-      const isExternal = _.get(post, 'is_external', false)
-      const href = getHref(_.get(post, 'slug', 'error'), isExternal)
-      const fadingStyle = {
-        opacity: this.state.selected === index ? '1' : '0',
-        zIndex: this.state.selected === index ? '1' : '0',
-        transition: 'opacity .5s linear',
-      }
-      return (
-        <FadeInFadeOut
-          key={_.get(post, 'id')}
-          isSelected={index === this.state.selected}
-        >
-          <Category>{_.get(post, 'categories[0].name', '')}</Category>
-          <Title style={fadingStyle}>
-            <TRLink href={href} redirect={isExternal}>
-              <TitleSpan>{_.get(post, 'title', '')}</TitleSpan>
-            </TRLink>
-          </Title>
-          <Description>{_.get(post, 'og_description', '')}</Description>
-        </FadeInFadeOut>
-      )
     })
 
     return (
-      <CarouselContainer mobileWidth={mobileWidth}>
-        <Swipeable
-          onSwipedRight={this.onSwipedRight}
-          onSwipedLeft={this.onSwipedLeft}
-          onSwiping={onSwiping}
-        >
-          <SectionName mobileWidth={mobileWidth}>
-            <span>{sectionStrings.editorPick}</span>
-          </SectionName>
-          <TextFrame>{textFrameContent}</TextFrame>
-          <MobileFlexSwipeable.FlexList
-            selected={this.state.selected}
-            mobileWidth={mobileWidth}
-          >
-            {flexItems}
-          </MobileFlexSwipeable.FlexList>
-        </Swipeable>
+      <CarouselContainer>
+        <SectionName>
+          <span>{sectionStrings.editorPick}</span>
+        </SectionName>
+        <MobileSwiperList>{flexItems}</MobileSwiperList>
       </CarouselContainer>
     )
   }