+++ packages/react-components/src/button/components/link.js
@@ -30,9 +30,18 @@ const LinkContainer = styled(Link)`
   text-decoration-line: ${props => style.decoration[props.type]};
   color: ${props => style.color[props.type]};
 
-  &:hover {
-    text-decoration-line: underline;
-  }
+  ${props =>
+    props.disabled
+      ? `
+    opacity: 0.5;
+    cursor: auto;
+  `
+      : `
+    &:hover {
+      text-decoration-line: underline;
+    }
+    -webkit-tap-highlight-color: transparent;
+  `}
 `
 
 const LinkButton = ({
@@ -41,6 +50,7 @@ const LinkButton = ({
   text = '',
   weight = P1.Weight.NORMAL,
   TextComponent = null,
+  disabled = false,
   ...props
 }) => {
   const textJSX = TextComponent ? (
@@ -50,7 +60,7 @@ const LinkButton = ({
   )
 
   return (
-    <LinkContainer type={type} {...link} {...props}>
+    <LinkContainer type={type} disabled={disabled} {...link} {...props}>
       {textJSX}
     </LinkContainer>
   )
@@ -61,6 +71,7 @@ LinkButton.propTypes = {
   text: PropTypes.string.isRequired,
   weight: P1.propTypes.weight,
   TextComponent: PropTypes.elementType,
+  disabled: PropTypes.bool,
 }
 LinkButton.Type = LinkType
 LinkButton.Weight = P1.Weight